+++ admin/src/router.tsx
@@ -4,6 +4,7 @@ import Layout from "./components/Layout";
 import Login from "./pages/Login";
 import Lottery from "./pages/Lottery";
 import LotteryWinner from "./pages/LotteryWinner";
+import LotteryWinnerList from "./pages/LotteryWinnerList";
 import Rush from "./pages/Rush";
 
 export const router = createBrowserRouter([
@@ -26,6 +27,10 @@ export const router = createBrowserRouter([
                     {
                         path: "winner",
                         element: <LotteryWinner />,
+                    },
+                    {
+                        path: "winner-list",
+                        element: <LotteryWinnerList />,
                     },
                 ],
             },
